modified:   lib/models/account_owner/more/transactions/wallet_balance.dart 	modified:   lib/services/account_owner/more/transactions/withdrawal_service.dart 	modified:   lib/views/account_owner/more/widget/settings/screen/settings_screen.dart 	modified:   lib/views/account_owner/more/widget/settings/widget/PIN_settings/change_PIN_screen.dart 	modified:   lib/views/account_owner/more/widget/settings/widget/PIN_settings/forgot_PIN_screen.dart 	modified:   lib/views/account_owner/more/widget/transactions/trx_screen/trx_dashboard.dart

+++ lib/models/account_owner/more/transactions/wallet_balance.dart
@@ -3,11 +3,11 @@ class WalletBalance{
     required this.wallet_balance,
     required this.has_wallet_pin,
   });
-  late final int wallet_balance;
+  late final double wallet_balance;
   late final bool has_wallet_pin;
   
   WalletBalance.fromJson(Map<String, dynamic> json,){
-    wallet_balance = json['wallet_balance'] ?? 0;
+    wallet_balance = json['wallet_balance'] ?? 0.0;
     has_wallet_pin = json['has_wallet_pin'] ?? false;
   }
 }
